// src/generated/model/studentResponseDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ITC Summer
 * ITCancún summer API
 * OpenAPI spec version: 1.0
 */
import type { StudentResponseDtoControlNumber } from './studentResponseDtoControlNumber';
import type { StudentResponseDtoPhone } from './studentResponseDtoPhone';
import type { UserProfileResponseDto } from './userProfileResponseDto';
import type { StudentResponseDtoCareer } from './studentResponseDtoCareer';

export interface StudentResponseDto {
  id_student: number;
  control_number: StudentResponseDtoControlNumber;
  phone: StudentResponseDtoPhone;
  user: UserProfileResponseDto;
  career: StudentResponseDtoCareer;
}